vtk backed volume store
this is an example of a volume stored (and thus singlefile) backend.
premise is that each unique parameter set maps to an integer index
and that index corresponds to a slice within the volume. The should
fix the issues of precision in filenames and enable better
compression. It will also let us use some slick VTK volume filters
in places.

TODO:
should be possible to store things other than images
split out vtk dependency into its own variant of SFS