Add Layers and Fields
Layers are for objects that are displayed alone in the store but can
be composed together in the viewer. Fields are color choices for each
layer. In explorers, layers are hooked up with a generated class
that knows how to turn on/off the display of each thing.

Also remove validate_parameter which was dead code for long ago.

TODO: make viewer compose them, enable them in paraview explorer, fix
sublayers.